Comit przed sprzedaza kompa

+++ src/Outputs/OutputData.hpp
@@ -4,10 +4,10 @@
 
 struct OutputData
 {
-    int8_t throttle = 0; //0-100
-    int8_t pitch = 0; //-100-100
-    int8_t roll = 0; //-100-100
-    int8_t yaw = 0; //-100-100
+    int throttle = 0; //0-100
+    int pitch = 0; //-100-100
+    int roll = 0; //-100-100
+    int yaw = 0; //-100-100
 
     bool isArmed = false; 
     bool isConnected = false;
